backend/requirements.txt
torch==2.1.2
transformers==4.36.2
datasets==2.16.1
peft==0.7.1
trl==0.7.10
bitsandbytes==0.42.0
accelerate==0.26.1

bandit==1.7.9
radon==6.0.1

pytest==8.0.0
pytest-cov==4.1.0
pynguin==0.30.0
mutmut==2.4.1

google-generativeai==0.5.4